Rename and restructure he code
