+++ assets/js/app.js
@@ -29,10 +29,10 @@ import {initialize as initSidebar} from './sidebar'
 import {initialize as initVersions} from './versions'
 import {initialize as initNightMode} from './night'
 import {initialize as initMakeup} from './makeup'
-import {initialize as initFocusMode} from './focus_mode'
 import {initialize as initKeyboardShortcuts} from './keyboard-shortcuts'
 import {initialize as initQuickSwitch} from './quick-switch'
-import {initialize as initPopovers} from './popovers'
+import {initialize as initTooltips} from './tooltips/tooltips'
+import {initialize as initHintsPage} from './tooltips/hints-page'
 
 window.$ = $
 
@@ -63,14 +63,14 @@ $(() => {
   hljs.registerLanguage('xml', xml)
 
   initNightMode()
-  initFocusMode()
   initSidebar()
   initVersions()
   initEvents()
   initMakeup()
   initKeyboardShortcuts()
   initQuickSwitch()
-  initFocusMode()
-  initPopovers()
+  initTooltips()
+  initHintsPage()
+
   hljs.initHighlighting()
 })
