+++ assets/less/app.less
@@ -25,8 +25,6 @@
 @import './night/keyboard-shortcuts';
 @import './night/popover';
 
-@import './focus-mode';
-
 @import './screen-reader';
 @import './print';
 @import './makeup';
