--- assets/less/focus-mode.less
@@ -1,40 +1,0 @@
-body.focus-mode {
-  &.sidebar-opened .content {
-    padding-left: 0;
-  }
-
-  button.sidebar-button {
-    display: none;
-  }
-
-  .sidebar {
-    display: none;
-  }
-
-  .content-inner {
-    display: block;
-    padding: 0;
-
-    .view-source,
-    .detail-link {
-      display: none;
-    }
-
-    .detail-header {
-      border-width: 0;
-      margin-top: 0;
-    }
-
-    .docstring {
-      margin: 1.2em;
-
-      > * {
-        display: none;
-      }
-
-      p:first-child {
-        display: block;
-      }
-    }
-  }
-}
